+++ packages/doc-sync/src/types.ts
@@ -1,5 +1,5 @@
 import { Response } from '@erebos/bzz'
-import { BzzFeed, FeedParams } from '@erebos/bzz-feed'
+import { BzzFeed, Feed, FeedParams } from '@erebos/bzz-feed'
 import { DataListReader, DataListWriter } from '@erebos/feed-list'
 import { Change, Doc } from 'automerge'
 
@@ -36,19 +36,19 @@ export interface DataPayload extends DataContent, ProtocolContent {}
 export interface MetaPayload extends MetaContent, ProtocolContent {}
 
 export interface DocFeeds {
-  data: FeedParams
-  meta: FeedParams
+  data: Feed
+  meta: Feed
 }
 
 export interface DocSerialized {
   docString: string
   dataFeed: FeedParams
-  metaFeed: FeedParams
+  metaFeed: Feed
 }
 
 export interface LoadDocReaderParams<B extends Bzz = Bzz> {
   bzz: B
-  feed: FeedParams // params for meta feed
+  feed: Feed // meta feed
 }
 
 export interface FromJSONDocReaderParams<B extends Bzz = Bzz>
@@ -59,30 +59,36 @@ export interface FromJSONDocReaderParams<B extends Bzz = Bzz>
 export interface DocReaderParams<T, B extends Bzz = Bzz> {
   bzz: B
   doc: Doc<T>
-  feed: FeedParams
+  feed: Feed
   list: DataListReader<DataPayload, B>
   time: number
 }
 
-export interface DocSubscriberParams<T, B extends Bzz = Bzz>
-  extends DocReaderParams<T, B> {
+interface SubscriberParams {
   pullInterval: number
 }
+
+export interface DocSubscriberParams<T, B extends Bzz = Bzz>
+  extends DocReaderParams<T, B>,
+    SubscriberParams {}
 
 export interface FromJSONDocSubscriberParams<B extends Bzz = Bzz>
-  extends FromJSONDocReaderParams<B> {
-  pullInterval: number
-}
+  extends FromJSONDocReaderParams<B>,
+    SubscriberParams {}
 
 export interface LoadDocSubscriberParams<B extends Bzz = Bzz>
-  extends LoadDocReaderParams<B> {
-  pullInterval: number
+  extends LoadDocReaderParams<B>,
+    SubscriberParams {}
+
+interface WriterParams {
+  signParams?: any
+  snapshotFrequency?: number
 }
 
-export interface CreateDocWriterParams<B extends Bzz = Bzz> {
+export interface CreateDocWriterParams<B extends Bzz = Bzz>
+  extends WriterParams {
   bzz: B
   feed: FeedFactoryParams // params used to create both data and meta feeds
-  snapshotFrequency?: number
 }
 
 export interface InitDocWriterParams<T, B extends Bzz = Bzz>
@@ -91,42 +97,40 @@ export interface InitDocWriterParams<T, B extends Bzz = Bzz>
 }
 
 export interface FromJSONDocWriterParams<B extends Bzz = Bzz>
-  extends FromJSONDocReaderParams<B> {
-  snapshotFrequency?: number
-}
+  extends FromJSONDocReaderParams<B>,
+    WriterParams {}
 
 export interface LoadDocWriterParams<B extends Bzz = Bzz>
-  extends LoadDocReaderParams<B> {
-  snapshotFrequency?: number
-}
+  extends LoadDocReaderParams<B>,
+    WriterParams {}
 
-export interface DocWriterParams<T, B extends Bzz = Bzz> {
+export interface DocWriterParams<T, B extends Bzz = Bzz> extends WriterParams {
   bzz: B
   doc: Doc<T>
-  feed: FeedParams
+  feed: Feed
   list: DataListWriter<DataPayload, B>
-  snapshotFrequency?: number
+}
+
+interface SynchronizerParams extends SubscriberParams {
+  pushInterval?: number
 }
 
 export interface InitDocSynchronizerParams<T, B extends Bzz = Bzz>
-  extends InitDocWriterParams<T, B> {
-  pullInterval: number
-  pushInterval?: number
-  sources?: Array<FeedParams>
+  extends InitDocWriterParams<T, B>,
+    SynchronizerParams {
+  sources?: Array<Feed>
 }
 
 export interface FromJSONDocSynchronizerParams<B extends Bzz = Bzz>
-  extends FromJSONDocWriterParams<B> {
-  pullInterval: number
-  pushInterval?: number
+  extends FromJSONDocWriterParams<B>,
+    SynchronizerParams {
   sources?: Array<DocSerialized>
 }
 
 export interface LoadDocSynchronizerParams<B extends Bzz = Bzz>
-  extends LoadDocWriterParams<B> {
-  pullInterval: number
-  pushInterval?: number
-  sources?: Array<FeedParams>
+  extends LoadDocWriterParams<B>,
+    SynchronizerParams {
+  sources?: Array<Feed>
 }
 
 export interface DocSynchronizerParams<T, B extends Bzz = Bzz>