+++ packages/doc-sync/types/DocReader.d.ts
@@ -1,4 +1,4 @@
-import { FeedParams } from '@erebos/bzz-feed';
+import { Feed } from '@erebos/bzz-feed';
 import { DataListReader } from '@erebos/feed-list';
 import { Doc } from 'automerge';
 import { BehaviorSubject } from 'rxjs';
@@ -7,12 +7,12 @@ export declare class DocReader<T, B extends Bzz = Bzz> extends BehaviorSubject<D
     static fromJSON<T, B extends Bzz = Bzz>(params: FromJSONDocReaderParams<B>): DocReader<T, B>;
     static load<T, B extends Bzz = Bzz>(params: LoadDocReaderParams<B>): Promise<DocReader<T, B>>;
     protected bzz: B;
-    protected feed: FeedParams;
+    protected feed: Feed;
     protected list: DataListReader<DataPayload, B>;
     protected time: number;
     private pullPromise;
     constructor(params: DocReaderParams<T, B>);
-    get metaFeed(): FeedParams;
+    get metaFeed(): Feed;
     protected write(doc: Doc<T>): boolean;
     protected load(): Promise<boolean>;
     pull(): Promise<boolean>;